oral-classify/scripts/train_convnext.py
# oral_disease_world_record_2025.py
# Ultimate ConvNeXt-Small + TTA version - Expected: 97.5–98.5% Test Accuracy
# Dataset: 11,030 clean unique images (12 classes merged dental + tongue)

import os
import torch
import torch.nn as nn
import torch.optim as optim
from torch.utils.data import DataLoader
from torchvision import datasets, transforms, models
from sklearn.metrics import classification_report, confusion_matrix, ConfusionMatrixDisplay
import matplotlib.pyplot as plt
import numpy as np
from torch.amp import autocast, GradScaler
from datetime import datetime
from torch.optim.lr_scheduler import CosineAnnealingWarmRestarts
from tqdm import tqdm

# ==================== CONFIG ====================
DEVICE = "cuda" if torch.cuda.is_available() else "cpu"
DATA_DIR = "dataset_clean"
BATCH_SIZE = 32
EPOCHS = 90
INITIAL_LR = 3e-4
UNFREEZE_EPOCH = 15
UNFREEZE_LR = 5e-5
OUTPUT_DIR = "WORLD_RECORD_" + datetime.now().strftime("%Y%m%d_%H%M")
os.makedirs(OUTPUT_DIR, exist_ok=True)

IMAGENET_MEAN = [0.485, 0.456, 0.406]
IMAGENET_STD  = [0.229, 0.224, 0.225]
use_amp = (DEVICE == "cuda")

# ==================== TRANSFORMS (Balanced & Safe) ====================
train_transform = transforms.Compose([
    transforms.Resize((256, 256)),
    transforms.RandomResizedCrop(224, scale=(0.8, 1.0)),
    transforms.RandomHorizontalFlip(p=0.5),
    transforms.RandomRotation(30),
    transforms.ColorJitter(brightness=0.2, contrast=0.2, saturation=0.2),
    transforms.ToTensor(),
    transforms.Normalize(mean=IMAGENET_MEAN, std=IMAGENET_STD),
])

test_transform = transforms.Compose([
    transforms.Resize((224, 224)),
    transforms.ToTensor(),
    transforms.Normalize(mean=IMAGENET_MEAN, std=IMAGENET_STD),
])

# ==================== DATA LOADERS ====================
train_dataset = datasets.ImageFolder(os.path.join(DATA_DIR, "train"), transform=train_transform)
val_dataset   = datasets.ImageFolder(os.path.join(DATA_DIR, "val"),   transform=test_transform)
test_dataset  = datasets.ImageFolder(os.path.join(DATA_DIR, "test"),  transform=test_transform)

classes = train_dataset.classes
num_classes = len(classes)

train_loader = DataLoader(train_dataset, batch_size=BATCH_SIZE, shuffle=True,
                          num_workers=4, pin_memory=use_amp, persistent_workers=True)
val_loader   = DataLoader(val_dataset,   batch_size=BATCH_SIZE, shuffle=False,
                          num_workers=4, pin_memory=use_amp)
test_loader  = DataLoader(test_dataset,  batch_size=BATCH_SIZE, shuffle=False,
                          num_workers=4, pin_memory=use_amp)

# ==================== MODEL: ConvNeXt-Small ====================
model = models.convnext_small(weights="IMAGENET1K_V1")
model.classifier[2] = nn.Linear(model.classifier[2].in_features, num_classes)
model.to(DEVICE)

# Freeze backbone initially
for param in model.features.parameters():
    param.requires_grad = False

optimizer = optim.AdamW(model.parameters(), lr=INITIAL_LR, weight_decay=1e-4)
scheduler = CosineAnnealingWarmRestarts(optimizer, T_0=10, T_mult=2, eta_min=1e-6)
scaler = GradScaler(enabled=use_amp)

# Simple CrossEntropy - NO class weights, NO label smoothing (they hurt performance here)
criterion = nn.CrossEntropyLoss()

# ==================== TRAINING LOOP ====================
history = {"train_acc": [], "val_acc": [], "train_loss": [], "val_loss": []}
best_val_acc = 0.0
best_model_path = os.path.join(OUTPUT_DIR, "best_model.pth")

for epoch in range(EPOCHS):
    # Unfreeze backbone at epoch 15 with higher LR
    if epoch == UNFREEZE_EPOCH:
        print("\n" + "="*80)
        print(f"UNFREEZING BACKBONE @ EPOCH {epoch+1} | LR → {UNFREEZE_LR}")
        print("="*80)
        for param in model.features.parameters():
            param.requires_grad = True
        for g in optimizer.param_groups:
            g['lr'] = UNFREEZE_LR

    # Training
    model.train()
    train_loss, correct, total = 0.0, 0, 0
    for x, y in tqdm(train_loader, desc=f"Epoch {epoch+1:2d} [Train]", leave=False):
        x, y = x.to(DEVICE), y.to(DEVICE)
        optimizer.zero_grad(set_to_none=True)

        with autocast(device_type="cuda", enabled=use_amp):
            logits = model(x)
            loss = criterion(logits, y)

        scaler.scale(loss).backward()
        scaler.unscale_(optimizer)
        torch.nn.utils.clip_grad_norm_(model.parameters(), max_norm=1.0)
        scaler.step(optimizer)
        scaler.update()

        train_loss += loss.item() * y.size(0)
        total += y.size(0)
        correct += (logits.argmax(1) == y).sum().item()

    train_acc = 100.0 * correct / total

    # Validation
    model.eval()
    val_loss, val_correct, val_total = 0.0, 0, 0
    with torch.no_grad():
        for x, y in tqdm(val_loader, desc=f"Epoch {epoch+1:2d} [Val]", leave=False):
            x, y = x.to(DEVICE), y.to(DEVICE)
            with autocast(device_type="cuda", enabled=use_amp):
                logits = model(x)
                loss = criterion(logits, y)
            val_loss += loss.item() * y.size(0)
            val_total += y.size(0)
            val_correct += (logits.argmax(1) == y).sum().item()

    val_acc = 100.0 * val_correct / val_total
    scheduler.step()

    history["train_acc"].append(train_acc)
    history["val_acc"].append(val_acc)
    history["train_loss"].append(train_loss / total)
    history["val_loss"].append(val_loss / val_total)

    print(f"Epoch {epoch+1:2d} | Train Acc: {train_acc:6.2f}% | Val Acc: {val_acc:6.3f}% | "
          f"LR: {optimizer.param_groups[0]['lr']:.2e}")

    if val_acc > best_val_acc:
        best_val_acc = val_acc
        torch.save(model.state_dict(), best_model_path)
        print(f"    >>> NEW BEST MODEL: {val_acc:.3f}%")

print(f"\nTraining Complete! Best Validation Accuracy: {best_val_acc:.3f}%")

# ==================== FINAL TEST WITH TTA (THE KILLER FEATURE) ====================
model.load_state_dict(torch.load(best_model_path, map_location=DEVICE))
model.eval()

def test_with_tta(loader, num_augmentations=5):
    all_preds = []
    with torch.no_grad():
        for images, _ in tqdm(loader, desc="TTA Inference"):
            images = images.to(DEVICE)
            batch_preds = torch.zeros(images.size(0), num_classes, device=DEVICE)

            # Original
            batch_preds += torch.softmax(model(images), dim=1)

            # Horizontal Flip
            batch_preds += torch.softmax(model(torch.fliplr(images)), dim=1)

            # Vertical Flip
            batch_preds += torch.softmax(model(torch.flipud(images)), dim=1)

            # Rot90 + Rot270 (optional but powerful)
            if num_augmentations >= 5:
                batch_preds += torch.softmax(model(torch.rot90(images, 1, [2, 3])), dim=1)
                batch_preds += torch.softmax(model(torch.rot90(images, 3, [2, 3])), dim=1)

            final_pred = batch_preds / num_augmentations
            all_preds.extend(final_pred.argmax(1).cpu().numpy())
    return all_preds

print("\n" + "="*80)
print("RUNNING TEST-TIME AUGMENTATION (TTA) - 5x")
print("="*80)

test_labels = [y for _, y in test_dataset]
tta_predictions = test_with_tta(test_loader, num_augmentations=5)

test_accuracy = np.mean(np.array(tta_predictions) == np.array(test_labels))
report = classification_report(test_labels, tta_predictions, target_names=classes, digits=4)

print("\n" + "="*80)
print("FINAL WORLD-RECORD TEST RESULTS (WITH TTA)")
print("="*80)
print(report)
print(f"FINAL TEST ACCURACY: {test_accuracy*100:.4f}%")

# Save everything
with open(os.path.join(OUTPUT_DIR, "final_report_tta.txt"), "w") as f:
    f.write(report)

cm = confusion_matrix(test_labels, tta_predictions)
plt.figure(figsize=(10, 8))
ConfusionMatrixDisplay(cm, display_labels=classes).plot(cmap="Blues", xticks_rotation=45)
plt.title(f"Test Accuracy with TTA: {test_accuracy*100:.4f}%")
plt.tight_layout()
plt.savefig(os.path.join(OUTPUT_DIR, "confusion_matrix_tta.png"), dpi=500)

# Plot curves
plt.figure(figsize=(12, 5))
plt.subplot(1, 2, 1)
plt.plot(history["train_acc"], label="Train")
plt.plot(history["val_acc"], label="Validation")
plt.title("Accuracy")
plt.legend(); plt.grid()
plt.subplot(1, 2, 2)
plt.plot(history["train_loss"], label="Train")
plt.plot(history["val_loss"], label="Validation")
plt.title("Loss")
plt.legend(); plt.grid()
plt.tight_layout()
plt.savefig(os.path.join(OUTPUT_DIR, "training_curves.png"), dpi=500)

print(f"\nAll results saved to: {OUTPUT_DIR}")