oral-classify/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.0.0
opencv-python>=4.8.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
tqdm>=4.65.0

# API dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6

# Model hub
timm>=0.9.0
